121b0ef Added new event when the active plot changes
c4ecfc0 Removed warning macro for no input
fc4c570 Merge topic 'fix-volume-rendering-int-overflow2' into master
e3618b8 Merge topic 'fix-clang-encoding-warnings' into master
ce3bd79 Merge topic 'amr-plane-cutter' into master
e72902d Merge topic 'move-test-4-modularization' into master
9f3b903 Merge branch 'release'
ee48a55 Renaming file to fix possible modularization conflict.
3af266f Merge topic 'delete-unused-tests' into master
3206aa3 Merge topic 'delete-duplicate-tests' into master
958fe0c Deleting duplicate test.
8b3d93b Merge branch 'fix-python-vtk-inputs-r-rel' into release
68b73f8 Merge branch 'example-python-module-suffix-rel' into release
e179bdb Deleting unused tests
ef41e28 Merge topic 'add-child-info-check-to-amr' into master
9cb34f5 Add a function to check whether there exists parent child information
...

+++ Views/vtkRenderView.cxx
@@ -430,7 +430,10 @@ void vtkRenderView::GenerateSelection(void* callData, vtkSelection* sel)
   if (this->SelectionMode == FRUSTUM)
     {
     // Do a frustum selection.
-    double displayRectangle[4] = {screenMinX, screenMinY, screenMaxX, screenMaxY};
+    double displayRectangle[4] = {static_cast<double>(screenMinX),
+                                  static_cast<double>(screenMinY),
+                                  static_cast<double>(screenMaxX),
+                                  static_cast<double>(screenMaxY)};
     vtkSmartPointer<vtkDoubleArray> frustcorners =
       vtkSmartPointer<vtkDoubleArray>::New();
     frustcorners->SetNumberOfComponents(4);
